LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/traverseTo-intercept-rejected-expected.txt


FAIL traverseTo() promise rejection with rejected intercept() assert_equals: expected "#1" but got ""
